requirements.txt
fastapi
uvicorn
python-dotenv
contractions
pyyaml
python-multipart
pytesseract
Pillow